+++ CHANGES.md
@@ -116,3 +116,6 @@
  - Configurable timeout for shared-stream HttpMulticaster (#138)
  - Added RSS Bot
  - Fixed #146, #144, #142
+ 
+## April 2021
+  - Fix for #165, @Display annotation to choose how fields are shown.